+++ src/components/orchestrator/OrchestratorWrapper.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from '@emotion/styled'
 
-import SubNavigation from './SubNavigation'
-import { getOrchestratorSubNavigation } from '../util/AppUtil'
+import SubNavigation from './../SubNavigation'
+import { getOrchestratorSubNavigation } from '../../util/AppUtil'
 import Orchestrator from './Orchestrator'
 
 const OrchestratorWrapper = () => {
